Reorganizado el proyecto, pasando a una carpeta de aplicacion
